requirements.txt
python-telegram-bot==20.8
google-generativeai==0.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
sqlalchemy==2.0.25
asyncpg==0.29.0
aiohttp==3.9.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.23.3
